% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jar_embed.R
\name{jar_embed}
\alias{jar_embed}
\title{Joint attractor reconstruction}
\usage{
jar_embed(x, y, m = 2, tau = 1, method = c("concat", "multivar"))
}
\arguments{
\item{x, y}{Numeric vectors}

\item{m}{Embedding dimension}

\item{tau}{Delay}

\item{method}{concat or multivar}
}
\value{
Joint embedded trajectory
}
\description{
Joint attractor reconstruction
}
