% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrqa_compute.R
\name{jrqa_compute}
\alias{jrqa_compute}
\title{Joint recurrence quantification analysis}
\usage{
jrqa_compute(x, y, m = 2, tau = 1, eps, theiler = 0)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{m}{Embedding dimension}

\item{tau}{Delay}

\item{eps}{Radius}

\item{theiler}{Theiler window}
}
\value{
List with joint recurrence matrix and metrics
}
\description{
Joint recurrence quantification analysis
}
