% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrqa_metrics.R
\name{jrqa_metrics}
\alias{jrqa_metrics}
\title{JRQA metrics from a joint recurrence matrix}
\usage{
jrqa_metrics(R, lmin = 2)
}
\arguments{
\item{R}{Logical recurrence matrix}

\item{lmin}{Minimum diagonal line length}
}
\value{
Named list of JRQA metrics
}
\description{
JRQA metrics from a joint recurrence matrix
}
